1
1

+++ src/main/java/cn/edu/buaa/scholarshipserver/dao/InstitutionDao.java
@@ -4,6 +4,6 @@ import cn.edu.buaa.scholarshipserver.es.Institution;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
 public interface InstitutionDao extends ElasticsearchRepository<Institution,Long> {
-        public Institution findByInstitutionId(Integer institutionId);
+        public Institution findByInstitutionId(Long institutionId);
         public Institution findByInstitutionName(String institutionName);
 }
